requirements.txt
mysql-connector-python~=8.0.28
pymongo~=4.0.1
flask~=1.1.2
requests~=2.26.0
selenium~=4.1.3
Flask-Cors==3.0.9
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
Werkzeug==1.0.1
dnspython

